feat: add basic WalletDirector
