+++ packages/core-web/src/FedimintWallet.ts
@@ -1,4 +1,4 @@
-import { TransportClient, WasmWorkerTransport } from './transport'
+import { TransportClient } from './transport'
 import {
   BalanceService,
   MintService,
@@ -7,14 +7,10 @@ import {
   RecoveryService,
   WalletService,
 } from './services'
-import { type LogLevel } from './utils/logger'
-import { FederationConfig, JSONValue, Transport } from './types'
 
 const DEFAULT_CLIENT_NAME = 'fm-default' as const
 
 export class FedimintWallet {
-  private _client: TransportClient
-
   public balance: BalanceService
   public mint: MintService
   public lightning: LightningService
@@ -41,11 +37,6 @@ export class FedimintWallet {
    * the first operation that requires them, reducing initial overhead at the cost
    * of a slight delay on the first operation.
    *
-   * @param {boolean} lazy - If true, delays Web Worker and WebAssembly initialization
-   *                         until needed. Default is false.
-   * @param {Transport} [transport] - Optional worker client instance. Provide your
-   *                         own to use a custom transport (e.g. React Native).
-   *
    * @example
    * // Create a wallet with immediate initialization
    * const wallet = new FedimintWallet();
@@ -57,32 +48,16 @@ export class FedimintWallet {
    * lazyWallet.initialize();
    * lazyWallet.open();
    */
-  constructor(
-    lazy: boolean = false,
-    transport: Transport = new WasmWorkerTransport(),
-  ) {
+  constructor(private _client: TransportClient) {
     this._openPromise = new Promise((resolve) => {
       this._resolveOpen = resolve
     })
-    this._client = new TransportClient(transport)
     this.mint = new MintService(this._client)
     this.lightning = new LightningService(this._client)
     this.balance = new BalanceService(this._client)
     this.federation = new FederationService(this._client)
     this.recovery = new RecoveryService(this._client)
     this.wallet = new WalletService(this._client)
-
-    this._client.logger.info('FedimintWallet instantiated')
-
-    if (!lazy) {
-      this.initialize()
-    }
-  }
-
-  async initialize() {
-    this._client.logger.info('Initializing TransportClient')
-    await this._client.initialize()
-    this._client.logger.info('TransportClient initialized')
   }
 
   async waitForOpen() {
@@ -91,7 +66,6 @@ export class FedimintWallet {
   }
 
   async open(clientName: string = DEFAULT_CLIENT_NAME) {
-    await this._client.initialize()
     // TODO: Determine if this should be safe or throw
     if (this._isOpen) throw new Error('The FedimintWallet is already open.')
     const { success } = await this._client.sendSingleMessage<{
@@ -108,7 +82,6 @@ export class FedimintWallet {
     inviteCode: string,
     clientName: string = DEFAULT_CLIENT_NAME,
   ) {
-    await this._client.initialize()
     // TODO: Determine if this should be safe or throw
     if (this._isOpen)
       throw new Error(
@@ -142,79 +115,5 @@ export class FedimintWallet {
 
   isOpen() {
     return this._isOpen
-  }
-
-  async previewFederation(inviteCode: string) {
-    const response = this._client.sendSingleMessage<{
-      config: FederationConfig
-      federation_id: string
-    }>('previewFederation', { inviteCode })
-    return response
-  }
-
-  /**
-   * Sets the log level for the library.
-   * @param level The desired log level ('DEBUG', 'INFO', 'WARN', 'ERROR', 'NONE').
-   */
-  setLogLevel(level: LogLevel) {
-    this._client.logger.setLevel(level)
-    this._client.logger.info(`Log level set to ${level}.`)
-  }
-
-  /**
-   * Parses a federation invite code and retrieves its details.
-   *
-   * This method sends the provided invite code to the TransportClient for parsing.
-   * The response includes the federation_id and url.
-   *
-   * @param {string} inviteCode - The invite code to be parsed.
-   * @returns {Promise<{ federation_id: string, url: string}>}
-   *          A promise that resolves to an object containing:
-   *          - `federation_id`: The id of the feder.
-   *          - `url`: One of the apipoints to connect to the federation
-   *
-   * @throws {Error} If the TransportClient encounters an issue during the parsing process.
-   *
-   * @example
-   * const inviteCode = "example-invite-code";
-   * const parsedCode = await wallet.parseInviteCode(inviteCode);
-   * console.log(parsedCode.federation_id, parsedCode.url);
-   */
-  async parseInviteCode(inviteCode: string) {
-    const response = await this._client.sendSingleMessage<{
-      type: string
-      data: JSONValue
-      requestId: number
-    }>('parseInviteCode', { inviteCode })
-    return response
-  }
-
-  /**
-   * Parses a BOLT11 Lightning invoice and retrieves its details.
-   *
-   * This method sends the provided invoice string to the TransportClient for parsing.
-   * The response includes details such as the amount, expiry, and memo.
-   *
-   * @param {string} invoiceStr - The BOLT11 invoice string to be parsed.
-   * @returns {Promise<{ amount: string, expiry: number, memo: string }>}
-   *          A promise that resolves to an object containing:
-   *          - `amount`: The amount specified in the invoice.
-   *          - `expiry`: The expiry time of the invoice in seconds.
-   *          - `memo`: A description or memo attached to the invoice.
-   *
-   * @throws {Error} If the TransportClient encounters an issue during the parsing process.
-   *
-   * @example
-   * const invoiceStr = "lnbc1...";
-   * const parsedInvoice = await wallet.parseBolt11Invoice(invoiceStr);
-   * console.log(parsedInvoice.amount, parsedInvoice.expiry, parsedInvoice.memo);
-   */
-  async parseBolt11Invoice(invoiceStr: string) {
-    const response = await this._client.sendSingleMessage<{
-      type: string
-      data: JSONValue
-      requestId: number
-    }>('parseBolt11Invoice', { invoiceStr })
-    return response
   }
 }
