+++ packages/core-web/src/index.ts
@@ -1,2 +1,2 @@
-export { FedimintWallet } from './FedimintWallet'
+export { WalletDirector } from './WalletDirector'
 export type * from './types'
